Edit README.md file to state current status of project
Note and Attribute section is added to tell user about the code and
remind the I have to add GUI in this project. Though it was simple but
had valuable lessons of javascript and writing code. For now its completed
and to to publicly available.TOP first javascript project Assignment Completed.